migrate confidential assets to separated package wip
