+++ confidential-assets/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@aptos-labs/confidential-assets",
+  "version": "0.1.0",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "description": "",
+  "dependencies": {
+    "@noble/curves": "^1.6.0",
+    "@noble/hashes": "^1.5.0",
+    "@aptos-labs/ts-sdk": "^1.38.0"
+  }
+}
