+++ confidential-assets/src/rangeProof.ts
