+++ confidential-assets/src/twistedElGamal.ts
@@ -3,9 +3,9 @@
 
 import { ed25519, RistrettoPoint } from "@noble/curves/ed25519";
 import { bytesToNumberLE } from "@noble/curves/abstract/utils";
-import { HexInput } from "../../types";
 import { H_RISTRETTO, RistPoint, TwistedEd25519PrivateKey, TwistedEd25519PublicKey } from "./twistedEd25519";
 import { ed25519GenRandom, ed25519modN } from "./utils";
+import { HexInput } from "@aptos-labs/ts-sdk";
 
 export interface DecryptionRange {
   start?: bigint;
