+++ confidential-assets/tests/units/checkIsNormalized.test.ts
