+++ confidential-assets/tests/units/checkIsRegistered.test.ts
