+++ confidential-assets/tests/units/confidentialProofs.test.ts
