+++ confidential-assets/tests/units/getAssetAuditor.test.ts
