+++ confidential-assets/tests/units/negativeNormalize.test.ts
