+++ confidential-assets/tests/units/negativeWithdraw.test.ts
