+++ confidential-assets/tests/units/normalize.test.ts
