+++ confidential-assets/tests/units/register.test.ts
