+++ confidential-assets/tests/units/safelyRotateKey.test.ts
