+++ confidential-assets/tests/units/transferWithAuditor.test.ts
