+++ package.json
@@ -52,13 +52,11 @@
   "dependencies": {
     "@aptos-labs/aptos-cli": "^1.0.2",
     "@aptos-labs/aptos-client": "^1.1.0",
-    "@distributedlab/aptos-wasm-bindings": "^0.3.16",
     "@noble/curves": "^1.6.0",
     "@noble/hashes": "^1.5.0",
     "@scure/bip32": "^1.4.0",
     "@scure/bip39": "^1.3.0",
     "@aptos-labs/script-composer-pack": "^0.0.9",
-    "comlink": "^4.4.2",
     "eventemitter3": "^5.0.1",
     "form-data": "^4.0.0",
     "js-base64": "^3.7.7",
