+++ src/api/aptos.ts
@@ -16,7 +16,6 @@ import { Table } from "./table";
 import { Keyless } from "./keyless";
 import { AptosObject } from "./object";
 import { AccountAbstraction } from "./account/abstraction";
-import { ConfidentialAsset } from "./confidentialAsset";
 
 /**
  * The main entry point for interacting with the Aptos APIs,
@@ -70,8 +69,6 @@ export class Aptos {
 
   readonly keyless: Keyless;
 
-  readonly confidentialAsset: ConfidentialAsset;
-
   readonly object: AptosObject;
 
   /**
@@ -110,7 +107,6 @@ export class Aptos {
     this.transaction = new Transaction(this.config);
     this.table = new Table(this.config);
     this.keyless = new Keyless(this.config);
-    this.confidentialAsset = new ConfidentialAsset(this.config);
     this.object = new AptosObject(this.config);
   }
 }
@@ -119,19 +115,18 @@ export class Aptos {
 // from the other classes will be recognized by typescript.
 export interface Aptos
   extends Account,
-  ANS,
-  Coin,
-  DigitalAsset,
-  Event,
-  Faucet,
-  FungibleAsset,
-  General,
-  Keyless,
-  Staking,
-  Table,
-  AptosObject,
-  ConfidentialAsset,
-  Omit<Transaction, "build" | "simulate" | "submit" | "batch"> { }
+    ANS,
+    Coin,
+    DigitalAsset,
+    Event,
+    Faucet,
+    FungibleAsset,
+    General,
+    Keyless,
+    Staking,
+    Table,
+    AptosObject,
+    Omit<Transaction, "build" | "simulate" | "submit" | "batch"> {}
 
 /**
 In TypeScript, we can’t inherit or extend from more than one class,
@@ -169,4 +164,3 @@ applyMixin(Aptos, Transaction, "transaction");
 applyMixin(Aptos, Table, "table");
 applyMixin(Aptos, Keyless, "keyless");
 applyMixin(Aptos, AptosObject, "object");
-applyMixin(Aptos, ConfidentialAsset, "object");
