--- src/core/crypto/confidential/index.ts
@@ -1,8 +1,0 @@
-// Copyright © Aptos Foundation
-// SPDX-License-Identifier: Apache-2.0
-
-export * from "./confidentialKeyRotation";
-export * from "./confidentialNormalization";
-export * from "./confidentialTransfer";
-export * from "./confidentialWithdraw";
-export * from "./confidentialAmount";
