+++ src/core/crypto/index.ts
@@ -15,9 +15,3 @@ export * from "./publicKey";
 export * from "./secp256k1";
 export * from "./signature";
 export * from "./singleKey";
-export * from "./twistedEd25519";
-export * from "./twistedElGamal";
-export * from "./confidential";
-export * from "./rangeProof";
-export * from "./kangaroo/kangarooRistretto";
-export * from "./kangaroo/tableMap";
